refactor: rename unauth in client
