+++ src/client/globals.rs
