+++ src/client/mod.rs
@@ -1,5 +1,5 @@
+mod globals;
 mod queries;
-mod unauth;
 mod user_login;
 
 use reqwest::{header, Client as GraphQLClient};
